add unittest and fix bugs
