+++ examples/pytorch/metapath2vec/reading_data.py
@@ -1,127 +1,96 @@
-import numpy as np
-import torch
-from torch.utils.data import Dataset
-from download import AminerDataset
-np.random.seed(12345)
-
-class DataReader:
-    NEGATIVE_TABLE_SIZE = 1e8
-
-    def __init__(self, dataset, min_count, care_type):
-
-        self.negatives = []
-        self.discards = []
-        self.negpos = 0
-        self.care_type = care_type
-        self.word2id = dict()
-        self.id2word = dict()
-        self.sentences_count = 0
-        self.token_count = 0
-        self.word_frequency = dict()
-        self.inputFileName = dataset.fn
-        self.read_words(min_count)
-        self.initTableNegatives()
-        self.initTableDiscards()
-
-    def read_words(self, min_count):
-        word_frequency = dict()
-        for line in open(self.inputFileName, encoding="ISO-8859-1"):
-            line = line.split()
-            if len(line) > 1:
-                self.sentences_count += 1
-                for word in line:
-                    if len(word) > 0:
-                        self.token_count += 1
-                        word_frequency[word] = word_frequency.get(word, 0) + 1
-
-                        if self.token_count % 1000000 == 0:
-                            print("Read " + str(int(self.token_count / 1000000)) + "M words.")
-
-        wid = 0
-        for w, c in word_frequency.items():
-            if c < min_count:
-                continue
-            self.word2id[w] = wid
-            self.id2word[wid] = w
-            self.word_frequency[wid] = c
-            wid += 1
-
-        self.word_count = len(self.word2id)
-        print("Total embeddings: " + str(len(self.word2id)))
-
-    def initTableDiscards(self):
-        # get a frequency table for sub-sampling. Note that the frequency is adjusted by
-        # sub-sampling tricks.
-        t = 0.0001
-        f = np.array(list(self.word_frequency.values())) / self.token_count
-        self.discards = np.sqrt(t / f) + (t / f)
-
-    def initTableNegatives(self):
-        # get a table for negative sampling, if word with index 2 appears twice, then 2 will be listed
-        # in the table twice.
-        pow_frequency = np.array(list(self.word_frequency.values())) ** 0.75
-        words_pow = sum(pow_frequency)
-        ratio = pow_frequency / words_pow
-        count = np.round(ratio * DataReader.NEGATIVE_TABLE_SIZE)
-        for wid, c in enumerate(count):
-            self.negatives += [wid] * int(c)
-        self.negatives = np.array(self.negatives)
-        np.random.shuffle(self.negatives)
-        self.sampling_prob = ratio
-
-    def getNegatives(self, target, size):  # TODO check equality with target
-        if self.care_type == 0:
-            response = self.negatives[self.negpos:self.negpos + size]
-            self.negpos = (self.negpos + size) % len(self.negatives)
-            if len(response) != size:
-                return np.concatenate((response, self.negatives[0:self.negpos]))
-        return response
+import dgl
 
 
-# -----------------------------------------------------------------------------------------------------------------
+class MetaPathGenerator:
+	def __init__(self):
+		self.id_author = dict()
+		self.id_conf = dict()
+		self.author_coauthorlist = dict()
+		self.conf_authorlist = dict()
+		self.author_conflist = dict()
+		self.paper_author = dict()
+		self.author_paper = dict()
+		self.conf_paper = dict()
+		self.paper_conf = dict()
 
-class Metapath2vecDataset(Dataset):
-    def __init__(self, data, window_size):
-        # read in data, window_size and input filename
-        self.data = data
-        self.window_size = window_size
-        self.input_file = open(data.inputFileName, encoding="ISO-8859-1")
+	def read_data(self, dirpath):
+		with open(dirpath + "/id_author.txt") as adictfile:
+			for line in adictfile:
+				toks = line.strip().split("\t")
+				if len(toks) == 2:
+					self.id_author[int(toks[0])] = toks[1].replace(" ", "")
 
-    def __len__(self):
-        # return the number of walks
-        return self.data.sentences_count
+		#print "#authors", len(self.id_author)
 
-    def __getitem__(self, idx):
-        # return the list of pairs (center, context, 5 negatives)
-        while True:
-            line = self.input_file.readline()
-            if not line:
-                self.input_file.seek(0, 0)
-                line = self.input_file.readline()
+		with open(dirpath + "/id_conf.txt") as cdictfile:
+			for line in cdictfile:
+				toks = line.strip().split("\t")
+				if len(toks) == 2:
+					newconf = toks[1].replace(" ", "")
+					self.id_conf[int(toks[0])] = newconf
 
-            if len(line) > 1:
-                words = line.split()
+		#print "#conf", len(self.id_conf)
 
-                if len(words) > 1:
-                    word_ids = [self.data.word2id[w] for w in words if
-                                w in self.data.word2id and np.random.rand() < self.data.discards[self.data.word2id[w]]]
+		with open(dirpath + "/paper_author.txt") as pafile:
+			for line in pafile:
+				toks = line.strip().split("\t")
+				if len(toks) == 2:
+					p, a = toks[0], toks[1]
+					if p not in self.paper_author:
+						self.paper_author[p] = []
+					self.paper_author[p].append(a)
+					if a not in self.author_paper:
+						self.author_paper[a] = []
+					self.author_paper[a].append(p)
 
-                    pair_catch = []
-                    for i, u in enumerate(word_ids):
-                        for j, v in enumerate(
-                                word_ids[max(i - self.window_size, 0):i + self.window_size]):
-                            assert u < self.data.word_count
-                            assert v < self.data.word_count
-                            if i == j:
-                                continue
-                            pair_catch.append((u, v, self.data.getNegatives(v,5)))
-                    return pair_catch
+		with open(dirpath + "/paper_conf.txt") as pcfile:
+			for line in pcfile:
+				toks = line.strip().split("\t")
+				if len(toks) == 2:
+					p, c = toks[0], toks[1]
+					self.paper_conf[p] = c
+					if c not in self.conf_paper:
+						self.conf_paper[c] = []
+					self.conf_paper[c].append(p)
+
+		sumpapersconf, sumauthorsconf = 0, 0
+		conf_authors = dict()
+		for conf in self.conf_paper:
+			papers = self.conf_paper[conf]
+			sumpapersconf += len(papers)
+			for paper in papers:
+				if paper in self.paper_author:
+					authors = self.paper_author[paper]
+					sumauthorsconf += len(authors)
+
+		print ("#confs  ", len(self.conf_paper))
+		print ("#papers ", sumpapersconf,  "#papers per conf ", sumpapersconf / len(self.conf_paper))
+		print ("#authors", sumauthorsconf, "#authors per conf", sumauthorsconf / len(self.conf_paper))
 
 
-    @staticmethod
-    def collate(batches):
-        all_u = [u for batch in batches for u, _, _ in batch if len(batch) > 0]
-        all_v = [v for batch in batches for _, v, _ in batch if len(batch) > 0]
-        all_neg_v = [neg_v for batch in batches for _, _, neg_v in batch if len(batch) > 0]
+	def generate_random_aca(self):
+		for conf in self.conf_paper:
+			self.conf_authorlist[conf] = []
+			for paper in self.conf_paper[conf]:
+				if paper not in self.paper_author: continue
+				for author in self.paper_author[paper]:
+					self.conf_authorlist[conf].append(author)
+					if author not in self.author_conflist:
+						self.author_conflist[author] = []
+					self.author_conflist[author].append(conf)
 
-        return torch.LongTensor(all_u), torch.LongTensor(all_v), torch.LongTensor(all_neg_v)
+		print ("author-conf list done")
+
+		con_list=[]
+		author_list=[]
+		for conf in self.conf_authorlist:
+			for a in self.conf_authorlist[conf]:
+				con_list.append(int(conf))
+				author_list.append(int(a))
+
+		print("list construct done")
+
+		hg = dgl.heterograph({
+			('conference', 'ca', 'author'): (con_list,author_list),
+			('author', 'ac', 'conference'): (author_list, con_list)})
+		return hg,self.id_conf,self.id_author
